Add support for java record serialization - move test code to google-cloud-firestore module by adding a test-jdk17 directory and a java17 profile, and eliminate google-cloud-firestore-jdk17 module
