--- google-cloud-firestore-jdk17/pom.xml
@@ -1,121 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>google-cloud-firestore-jdk17</artifactId>
-  <version>3.8.3-SNAPSHOT</version><!-- {x-version-update:google-cloud-firestore:current} -->
-  <packaging>jar</packaging>
-  <name>Google Cloud Firestore JDK17 extension</name>
-  <url>https://github.com/googleapis/java-firestore</url>
-  <description>
-    Java idiomatic client for Google Cloud Firestore.
-  </description>
-  <parent>
-    <groupId>com.google.cloud</groupId>
-    <artifactId>google-cloud-firestore-parent</artifactId>
-    <version>3.8.3-SNAPSHOT</version><!-- {x-version-update:google-cloud-firestore:current} -->
-  </parent>
-  <properties>
-    <site.installationModule>google-cloud-firestore-jdk17</site.installationModule>
-    <maven.compiler.target>17</maven.compiler.target>
-    <maven.compiler.source>17</maven.compiler.source>
-  </properties>
-
-  <dependencies>
-    <!-- Test dependencies -->
-    <dependency>
-      <groupId>javax.annotation</groupId>
-      <artifactId>javax.annotation-api</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.api</groupId>
-      <artifactId>api-common</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.cloud</groupId>
-      <artifactId>google-cloud-core</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.cloud</groupId>
-      <artifactId>google-cloud-firestore</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.api.grpc</groupId>
-      <artifactId>proto-google-cloud-firestore-v1</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.code.gson</groupId>
-      <artifactId>gson</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-core</artifactId>
-      <version>2.25.0</version>
-      <scope>test</scope>
-    </dependency>
-
-    <!-- Need testing utility classes for generated gRPC clients tests -->
-    <dependency>
-      <groupId>com.google.api</groupId>
-      <artifactId>gax</artifactId>
-      <classifier>testlib</classifier>
-      <scope>test</scope>
-    </dependency>
-</dependencies>
-
-  <reporting>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-javadoc-plugin</artifactId>
-        <version>3.5.0</version>
-        <configuration>
-          <show>protected</show>
-          <nohelp>true</nohelp>
-          <doclint>none</doclint>
-          <splitindex>true</splitindex>
-        </configuration>
-      </plugin>
-    </plugins>
-  </reporting>
-
-  <build>
-    <pluginManagement>
-      <plugins>
-        <plugin>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>3.10.1</version>
-          <configuration>
-            <source>17</source>
-            <target>17</target>
-            <encoding>UTF-8</encoding>
-            <compilerArgument>-Xlint:unchecked</compilerArgument>
-            <compilerArgument>-Xlint:deprecation</compilerArgument>
-            <showDeprecation>true</showDeprecation>
-          </configuration>
-        </plugin>
-      </plugins>
-    </pluginManagement>
-    <plugins>
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>flatten-maven-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-</project>
