+++ google-cloud-firestore/src/test-jdk17/java/com/google/cloud/firestore/RecordMapperTest.java
@@ -16,6 +16,18 @@
 
 package com.google.cloud.firestore;
 
+import com.google.cloud.firestore.annotation.DocumentId;
+import com.google.cloud.firestore.annotation.PropertyName;
+import com.google.cloud.firestore.annotation.ThrowOnExtraProperties;
+import com.google.cloud.firestore.spi.v1.FirestoreRpc;
+import com.google.common.collect.ImmutableList;
+import com.google.firestore.v1.DatabaseRootName;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mockito;
+import org.mockito.Spy;
+import org.mockito.junit.MockitoJUnitRunner;
+
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -28,22 +40,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import com.google.cloud.firestore.annotation.DocumentId;
-import com.google.cloud.firestore.annotation.PropertyName;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mockito;
-import org.mockito.Spy;
-
-import com.google.cloud.firestore.annotation.ThrowOnExtraProperties;
-import com.google.cloud.firestore.spi.v1.FirestoreRpc;
-import com.google.common.collect.ImmutableList;
-import com.google.firestore.v1.DatabaseRootName;
-import org.mockito.junit.MockitoJUnitRunner;
-
-import static com.google.cloud.firestore.jdk17.LocalFirestoreHelper.fromSingleQuotedString;
-import static com.google.cloud.firestore.jdk17.LocalFirestoreHelper.mapAnyType;
-import static org.junit.Assert.*;
+import static com.google.cloud.firestore.record.LocalFirestoreHelper.fromSingleQuotedString;
+import static com.google.cloud.firestore.record.LocalFirestoreHelper.mapAnyType;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThrows;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 @SuppressWarnings({"unused", "WeakerAccess", "SpellCheckingInspection"})
 @RunWith(MockitoJUnitRunner.class)
