+++ google-cloud-firestore/src/test-jdk17/java/com/google/cloud/firestore/record/LocalFirestoreHelper.java
@@ -14,24 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.firestore.jdk17;
-
-import java.math.BigInteger;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.TimeUnit;
-
-import javax.annotation.Nullable;
-
-import org.mockito.stubbing.Answer;
+package com.google.cloud.firestore.record;
 
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutures;
@@ -56,6 +39,21 @@ import com.google.gson.reflect.TypeToken;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.NullValue;
 import com.google.type.LatLng;
+import org.mockito.stubbing.Answer;
+
+import javax.annotation.Nullable;
+import java.math.BigInteger;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
