+++ pom.xml
@@ -216,7 +216,6 @@
     <module>proto-google-cloud-firestore-bundle-v1</module>
     <module>google-cloud-firestore-admin</module>
     <module>google-cloud-firestore-bom</module>
-    <module>google-cloud-firestore-jdk17</module>
   </modules>
 
   <reporting>
